[Refactor] 관리자, 회원 임시매장 api 분리
